refactor: split pipeline models into per-architecture files
Move WinMLEncoderDecoderModel + EncoderDecoderInputGenerator to
models/hf/encoder_decoder.py, DecoderOnlyInputGenerator to
models/winml/decoder_only.py. Colocate WinMLT5Model in t5.py and
WinMLQwen3Model in qwen.py alongside their export configs.

Rename seq2seq.py to pipeline_model.py (now only WinMLPipelineModel
+ registry). Add sub_model_kwargs to from_pretrained for per-component
shape_config. Generators accept max_cache_len/seq_len kwargs as
overrides from shape_config.
