+++ src/winml/modelkit/models/winml/decoder_only.py
@@ -47,7 +47,7 @@ How it works:
    rather than trimming to the last token.  On subsequent calls with a
    populated ``StaticCache``, we trim to the last token as usual.
 
-Design principles (same as seq2seq.py):
+Design principles (same as pipeline_model.py):
 
 - ONNX I/O names and shapes are read from ``io_config``, never hardcoded.
 - Inputs smaller than ONNX expected shape are zero-padded via ``_pad_inputs``.
@@ -57,30 +57,101 @@ Design principles (same as seq2seq.py):
 from __future__ import annotations
 
 import logging
-from typing import TYPE_CHECKING, Any, ClassVar
+from typing import TYPE_CHECKING, Any
 
 import torch
+from optimum.utils.input_generators import DummyInputGenerator
 from transformers import Cache, StaticCache
 from transformers.generation.utils import GenerationMixin
 from transformers.modeling_outputs import CausalLMOutputWithPast
 
-from . import register_specialization
-from .seq2seq import WinMLEncoderDecoderModel, WinMLPipelineModel, register_pipeline_model
+from .pipeline_model import WinMLPipelineModel
 
 
-# Reuse the static pad helper from WinMLEncoderDecoderModel
-_pad_inputs = WinMLEncoderDecoderModel._pad_inputs
-
-# Sub-models must use GenericTask (raw ONNX outputs) — task-specific
-# wrappers like WinMLModelForFeatureExtraction would discard KV outputs.
-register_specialization("qwen3", "feature-extraction", "WinMLModelForGenericTask")
-register_specialization("qwen3", "text-generation", "WinMLModelForGenericTask")
+_pad_inputs = WinMLPipelineModel._pad_inputs
 
 
 if TYPE_CHECKING:
     from transformers import PretrainedConfig
 
 logger = logging.getLogger(__name__)
+
+
+# =========================================================================
+# DecoderOnlyInputGenerator — shared dummy input generator
+# =========================================================================
+
+
+class DecoderOnlyInputGenerator(DummyInputGenerator):
+    """Generates base inputs for decoder-only models with static KV cache.
+
+    Produces ``input_ids``, ``attention_mask``, ``position_ids``, and
+    ``cache_position``.  Reads ``vocab_size``, ``max_cache_len``, and
+    ``seq_len`` from the ``NormalizedConfig``.
+
+    ``seq_len`` controls the input token count and is read from
+    ``normalized_config.seq_len`` (falls back to ``_default_seq_len``).
+    Subclasses override the default for prefill vs generation:
+
+    - ``DecoderOnlyPrefillInputGenerator``: ``_default_seq_len = 64``
+    - ``DecoderOnlyInputGenerator`` (base / gen): ``_default_seq_len = 1``
+
+    To override at config time, set ``config.seq_len = N`` on the HF config.
+    """
+
+    SUPPORTED_INPUT_NAMES = (
+        "input_ids",
+        "attention_mask",
+        "position_ids",
+        "cache_position",
+    )
+
+    _default_seq_len: int = 1
+
+    def __init__(
+        self,
+        task: str,
+        normalized_config: Any,
+        batch_size: int = 1,
+        seq_len: int | None = None,
+        max_cache_len: int | None = None,
+        **kwargs: Any,
+    ) -> None:
+        self.batch_size = batch_size
+        self.vocab_size = normalized_config.vocab_size
+        self.max_cache_len = max_cache_len or normalized_config.max_cache_len
+        self.seq_len: int = seq_len or getattr(normalized_config, "seq_len", self._default_seq_len)
+
+    def generate(
+        self,
+        input_name: str,
+        framework: str = "pt",
+        int_dtype: str = "int64",
+        float_dtype: str = "fp32",
+    ) -> torch.Tensor:
+        """Generate a dummy tensor for the given input name."""
+        if input_name == "input_ids":
+            return self.random_int_tensor(
+                (self.batch_size, self.seq_len),
+                max_value=self.vocab_size,
+                framework=framework,
+                dtype=int_dtype,
+            )
+        if input_name == "attention_mask":
+            mask = torch.zeros(self.batch_size, self.max_cache_len, dtype=torch.int64)
+            mask[:, : self.seq_len] = 1
+            return mask
+        if input_name == "position_ids":
+            return torch.arange(self.seq_len, dtype=torch.int64).unsqueeze(0)
+        if input_name == "cache_position":
+            return torch.arange(self.seq_len, dtype=torch.int64)
+        raise ValueError(f"Unknown input: {input_name}")
+
+
+class DecoderOnlyPrefillInputGenerator(DecoderOnlyInputGenerator):
+    """Prefill variant with ``_default_seq_len = 64``."""
+
+    _default_seq_len: int = 64
 
 
 # =========================================================================
@@ -302,42 +373,3 @@ class WinMLDecoderOnlyModel(WinMLPipelineModel, GenerationMixin):
             )
 
         return outputs["logits"]
-
-
-# =========================================================================
-# WinMLQwen3Model — Qwen3 tasks + generation config
-# =========================================================================
-
-
-@register_pipeline_model("qwen3", "text-generation")
-class WinMLQwen3Model(WinMLDecoderOnlyModel):
-    """Qwen3 decoder-only model.
-
-    Declares Qwen3 sub-component tasks and generation config defaults.
-    All forward/cache logic lives in ``WinMLDecoderOnlyModel``.
-    """
-
-    _SUB_MODEL_CONFIG: ClassVar[dict[str, str]] = {
-        "decoder_prefill": "feature-extraction",
-        "decoder_gen": "text-generation",
-    }
-
-    @property
-    def generation_config(self):  # noqa: D102
-        if not hasattr(self, "_generation_config"):
-            from transformers import GenerationConfig
-
-            gc_kw: dict[str, Any] = {}
-            for attr in ("bos_token_id", "eos_token_id", "pad_token_id"):
-                val = getattr(self.config, attr, None)
-                if val is not None:
-                    gc_kw[attr] = val
-            gc_kw.setdefault("max_new_tokens", self._max_cache_len - self._prefill_seq_len)
-            gc_kw.setdefault("num_beams", 1)
-            gc_kw.setdefault("do_sample", False)
-            self._generation_config = GenerationConfig(**gc_kw)
-        return self._generation_config
-
-    @generation_config.setter
-    def generation_config(self, value: Any) -> None:
-        self._generation_config = value
